purchase, usage, inventory, appln.prop

+++ src/main/java/com/inventory/system/inventory/dto/InventoryResponse.java
@@ -1,13 +1,15 @@
 package com.inventory.system.inventory.dto;
 
+import java.math.BigDecimal;
+
 public class InventoryResponse {
 	private Long itemId;
 	private String category;
 	private String itemName;
 	private String units;
-	private Double quantity;
-	private Double cost;
-	private Double total;
+	private BigDecimal quantity;
+	private BigDecimal cost;
+	private BigDecimal total;
 	private Integer minStock;
 
 	public String getCategory() {
@@ -34,27 +36,27 @@ public class InventoryResponse {
 		this.units = units;
 	}
 
-	public Double getQuantity() {
+	public BigDecimal getQuantity() {
 		return quantity;
 	}
 
-	public void setQuantity(Double quantity) {
+	public void setQuantity(BigDecimal quantity) {
 		this.quantity = quantity;
 	}
 
-	public Double getCost() {
+	public BigDecimal getCost() {
 		return cost;
 	}
 
-	public void setCost(Double cost) {
+	public void setCost(BigDecimal cost) {
 		this.cost = cost;
 	}
 
-	public Double getTotal() {
+	public BigDecimal getTotal() {
 		return total;
 	}
 
-	public void setTotal(Double total) {
+	public void setTotal(BigDecimal total) {
 		this.total = total;
 	}
